debugging COM SPI FIFO
COM SPI FIFO is supposed to be able to queue up COM requests
without loss should the EC become busy.

However, it appears this is not working. Incoming packets seem
to be getting lost if the CPU is busy and can't service
the FIFO before the next block is written.

This warrants further investigation, probably need to
build a simulation test bench to dig into more.