update keyboard wake-up chord
EVT required awful hacks because of interference from
the ESD network on the FPGA.

now the keyboard sense pins are fully isolated in standby,
so the scan polarity can now be positive (e.g. detect
high signal to indicate key pressed). This is desireable
because all the other pins connected to the FPGA appear as
if they are wired to ground when tested.

The main hack is the sense pin is pre-driven. Hardware
pull-down resistors would get rid of this step, but seems
to be reliable enough.