add FFI bindings for the LSM6DS3 inertial module
Unfortunately after many hours of work, this FFI is pretty
much a steaming pile of shit. I'm able to get it to report
the ID of the device, but for some reason merely compiling in
FFI calls -- and not calling them -- is enough to cause
the CPU to lock up. Can't even attach the debugger. I'm guessing
something got broken in the binding generation somewhere that's
causing e.g. function pointers to mis-asign.

A lot of futzing around seem to show that if I included
more than some critical number or combination of API calls,
the CPU will stop booting. I'm not sure what would cause this.

However, the main purpose of this exercise was hardware bringup.
To that end, I just skipped using the FFI provided by the vendor
and jammed a set of reasonable-looking hard-coded hex values
into the gyroscope, and now I can read back acceleration
values on X/Y/Z from the SoC.

So, the accelerometer works well enough to hit MVP. Not
tested are the automatic wake-on-tap functions (so like,
maybe the interrupt line isn't wired up right or something and
I wouldn't know it). That being said, the main purpose of
the unit can be fulfilled with the validation as is, that is,
detecting if the device has been moved when it's not supposed
to be moved.