revert capacity setting, make glitch detection more robust
GG seems to work better with native capacity -- user manual
is not telling us something important we should know about
resetting the capacity manually (seems like additional parameters
should be set). However, with the default capacity it seems to
accurately infer the actual battery capacity.

There is a glitch that happens on the voltage detection when
the battery is inserted or removed. This could be due to
improper locking of the I2C bus between shared processes, need
to look into it more!