fix malloc bugs
the original mallocator was broken

now we have a stupid mallocator that just doles out one
of several 4kiB segments to requests and frees them as they
are done.

Works as long as nobody wants more than 4kiB and/or there are
not more than a 8 outstanding requests.

In practice, so far I've only seen a couple hundred bytes max
and two concurrent mallocs; looks like the biggest malloc
would be the size of wifi frame plus a few bytes, which is
significantly below the 4kiB limit.