refactor power management so the SoC is largely in charge of its own power
previously, we were echoing our the value on the power state
back into the soc power on signal, which makes a race condition
for shutting things down.

Now, we only assert the SOC power on signal when we've detected
a charger is present, and once the SOC responds with anything,
we de-assert that line, and allow the SOC to kill itself "at will".