Use flow control for UART debug printing
This uses the TXFULL and TXEMPTY UART CSRs as an adaptive flow
control heuristic to mute debug printing when the EC guesses that
no wishbone-bridge serial tunnel is connected.

This change also avoids rapid polling of the wishbone bus. Reducing
wishbone bus activity seems to make wishbone-tool connections more
reliable and the firmware generally more stable. Running this
firmware without an active wishbone-tool link does not appear to
interfere with COM bus activity (Xous seems have no problem talking
to the gas gauge).

This code allows for reliable serial debug when the gateware is
built with debugonly=True. But it still gives an intermittent
wishbone-tool link when the gateware is built with debugonly=False.
In that case, some of the serial debug characters get lost, but the
connection mostly works.

Signed-off-by: samblenny