Overhaul logging and prune unused code
This commit is about cleaning up hal_wf200.rs to prepare for
attempting to add lwIP.

My initial goal was to prune unnecessary code and run rustfmt on
hal_wf200.rs. But, along the way, I hit a problem that needed trace
level logging of COM bus verbs. So, I also re-worked the logging.
The problem turned out to be a timing thing in the gateware. The
workaround fix was using `--seed=1` with betrusted_ec.py.

Changes:
- rustfmt sw/src/hal_wf200.rs
- Prune unused functions and blocks of commented-out code
- Prune code specific to WF200 AP mode (we only need STA)
- Significantly rework debug logging including:
  + Add a log level filtering mechanism
  + Shorten some strings, remove others
  + Try to make log messages easier to cross reference with docs
  + Update all 3 copies of debug.rs
- Remove wf200_get_rx_stats_raw() (unsure... was this needed?)
- Re-order COM verb handlers so related verbs are grouped together
- Replace `.unwrap()` calls with more efficient alternatives

Signed-off-by: samblenny