Restore TUSB320LAI revision debug log message
In an earlier commit, I removed a chip revision debug print buried
in sw/betrusted-hal/src/api_tusb320.rs because it caused a circular
dependency when I merged the previous three debug.rs modules into
a single crate. This restores the chip revision debug log in a way
that does not cause a dependency loop.

Signed-off-by: samblenny