Increase kernel max size from 56KiB to 76KiB
This tells xtask and the loader to allow an additional 20KiB for
EC kernel code. The current maximum possible code size is 96KiB.
The old kernel size limit of 56KiB allowed a 40KiB reserve for
network code. This change unlocks half of that reserve.

The 96KiB limit could potentially be increased if necessary by
reducing the RAM reserved for COM bus buffers.

Signed-off-by: samblenny