Implement a new IMU driver to replace gyro_*
This is a minimal LSM6DS3 IMU driver to do tap detection and XYZ
raw acceleration readout. I implemented this from scratch based on
the register documentation in STM's datasheet and app note AN5130.

The old gyro nomenclature should probably be deprecated as the chip
is an IMU with both an accelerometer and gyroscope. And, we're
using the accelerometer but not the gyroscope.

Tap detection seems to work okay, but the debouncing is fragile and
the IMU tap detection parameters could probably be tuned better.
Current debounce timer is clocked by iterations through the main
event loop, so tweaking unrelated things might break it. Not ideal.

Current kernel size: 56812 bytes

Signed-off-by: samblenny