Add UART RX function with wake sequence check
This adds debug UART RX polling to the EC event loop, but the UART
starts in bypass mode. After reset, incoming characters initially
get ignored. A state machine watches for the wake sequence of
"AT\n" or "AT\r". After the wake sequence is received, the RX
bypass gets turned off, and subsequent received chars get passed
through to the event loop.

The point of this is to provide a mechanism to trigger test packets
such as sending an ARP request or a DHCP request.

Signed-off-by: samblenny