Verify UDP header checksums
This adds an RFC 1071 one's complement UDP header checksum checker
with a counter for packets that fail validation.

Being able to calculate IP and UDP header checksums gets me closer
to generating valid DHCP request packets, which is the point of
this current quest. Next steps are at the DHCP protocol layer.

Signed-off-by: samblenny