Improve DHCP options parsing and debug logging
This steps through the DHCP options block, parsing and logging
each tag marking the start of an option field, along with its
length and data. Some of the most common tags are labeled with
their names. Lengths and data are logged as hex.

The point of better option parsing and logging is to prepare for
implementing a client to do DHCP discover and renew requests.
There are many DHCP options, and they can be packed in any order,
according to the server's choice. So, we need a solid parser to
interpret whatever the servers decide to reply with.

Signed-off-by: samblenny