Refactor randomizing for DHCP state machine
This shuffles around code involved in building a DHCPDISCOVER
packet so that the DhcpClient state machine struct can keep track
of xid and hostname. There is now an option to randomize the xid
and hostname to start a new transaction from the INIT state. But,
otherwise, the xid and hostname will be remembered and reused.

Signed-off-by: samblenny