Handle DHCPREQUEST and DHCPACK
This gets the DHCP client state machine working well enough to send
a valid DHCPREQUEST based on options parsed from an OFFER and to
switch to the BOUND state in response to a DHCPACK.

For now, none of the lease timers, exponential backoff retries, or
ARP resolution works. This just lets you manually step the state
machine clock to move through Init, Selecting, Requesting, Bound,
and so on. The result is your hostname and IP showing up in your
wifi router's list of connected devices.

Signed-off-by: samblenny