Simplify get_time_ms()
Before this change, get_time_ms() was fetching the whole 64-bit ms
ticktimer, dividing it by 1, then casting the result to u32. All
that amounted to just fetching the low word of the timer, but the
intermediate work compiled in an extra 279 bytes of kernel code.

This change skips all the intermediate stuff.

CAUTION: The 32-bit low word of the timer will roll over at 49.7
days. Callers currently need to handle that on their own. Would
probably be a good idea either switch everything to `u64` (slow),
or to make a struct for working with `u32` elapsed time intervals
that provides a function to check for rollover properly.

Signed-off-by: samblenny