Re-work delay_ms() to use full 40 bits of timer
The old delay_ms() was only using the bottom 32 bits of the
hardware ms timer. And, it was doing math in a way that would
panic for a delay interval that crossed the 49.7 day 32-bit timer
overflow. This change improves the timer to avoid panicking and to
use the full 40 timer bits.

This is a potentially RISKY change which should be reviewed with
suspicion. I think it's right. I hope it's right. But, this has
the potential to mess up anything with tight timing requirements.

Signed-off-by: samblenny