tried to do a `safe` implemenation of pkt_buf using inner mutability
unfortunately, because the C FFI function responsible for writing
packets into the buffer has to be able to side-effect the pkt_buf
from a callback, and you can't pass a reference to the object
to track ownership because of the structure of the C FFI, I think
there is no other way to do this than to use a whole bunch of unsafe
operations on a static mut object. :/

This checkpoint allows me to go back and try it again if I find
a better way to do this, but...after trying to chase out the borrow
checker it eventually came to the same conclusion, that we can't
reason about the safety of the C FFI callback function that
ultimately originates the packet data.