Add COM bus net bridge bypass AT command
This adds an AT command on the EC UART so that it's possible to
disconnect the COM bus net bridge. The disconnect works by not
sending RX packet interrupts and dropping COM bus TX packets
rather than sending them to the WF200.

Default state is COM bus net bridge *enabled*, so smoltcp will be
plumbed up normally unless you use the AT command to disconnect it.

This also removes most of the verbose debug printing of RX packets.
So, when the COM bus net bridge is disabled, the EC UART debug log
is much quieter. The idea is that you can use `tcpdump -X` on a
test LAN if you want hex dumps, and you can check the AT command
for network diagnostic stats counters if you want to see how many
of which types of packets were received by the EC.

Signed-off-by: samblenny