Use countdown timers for DHCP Renew & Rebind
This adds a function to the net crate's Countdown timer struct for
setting a timer interval in u32 seconds. That function hopefully
handles all the u64 math for the conversion to an interval in ms
that works with the 40-bit ms timer. I have not tested this enough
to have high confidence that it's correct, but so far it seems
pretty good.

Also, this changes the DHCP state machine clocking function to use
the new countdown timer features for setting the T1 and T2 renew
and rebind timers for 0.5 and 0.875 of the lease interval, as
suggested by the DHCP RFC.

I've tested this by watching tcpdump and the EC UART debug log
while manually disconnecting the Ethernet cable for my test LAN
DHCP server to trigger different code paths. As far as I can tell,
it works. But, I suspect the packets I'm sending may not fully
comply with all the DHCP RFC requirements for Request packets.

Signed-off-by: samblenny