Fix DHCP Renew & Rebind rebind timing
According to RFC 2131, renew and rebind should use retry delays of
half the remaining time to T2 and lease expiration, respectively,
down to a minimum of 60 seconds.

This just uses 60 second retries to avoid the hassle of plumbing
up the half-the-remaining-time calculations. I figure since this is
taking a little liberty with an RFC "SHOULD" requirement, rather
than a "MUST" requirement, this is acceptable.

Signed-off-by: samblenny