clear the SSID scan array after it is read
Instead of treating the SSID list as the accumulated results
of scanning, it's just a temp buffer that holds result until
a call happens to read all the results out. This removes the
chance of limit cycles causing some APS to disappear
due to the re-use of buffer entries based on environmental factors.